Spec2Struc.jl

Efficient statistical mechanical algorithms to generate effective potentials or configurations from pair correlation functions or structure factors.

Package Features

Torquato-Wang algorithm (Torquato and Wang, PRE 2020)

Iterative Boltzmann Inversion (Soper, Chem. Phys. 1996)

Reverse Monte-Carlo (McGreevy, J. Phys. Cond. Matter 2001)

Function Documentation

Spec2Struc.optim_parametrized_pot — Function

optim_parametrized_pot(my_params, pot, dim, ρ, targ_g2, targ_s; 
    large_r_grid = missing, n::Int = 600, bin_size::Float64 = 0.05, r_range::Float64 = 10, k_range::Float64 = 10,
    g2_weight_range::Float64 = 2, s_weight_range::Float64 = 4, 
    n_threads::Int = 15, configs_per_box::Int = 10, Ψ_tol::Float64 = 0.005, show_pb::Bool = true, test::Bool = false)

Using the Torquato-Wang algorithm to perform iterative optimization of potential parameters to match target pair correlation function and structure factor.

Arguments

my_params::Vector{Float64}: Vector of initial potential parameters.

pot::Function: Potential function that calculates the interaction potential between particles.

dim::Int: Dimension of the system.

ρ::Float64: Density of the system.

targ_g2::Function: Function representing the target pair correlation function. Accepts a distance value r and returns the target g2 value at that distance.

targ_s::Function: Function representing the target structure factor. Accepts a wave vector k and returns the target S value at that wave vector.

Keyword Arguments (all are optional)

large_r_grid::Missing: Large-r grid for computation of long-ranged potentials. Default value is missing.

n::Int: Number of boxes for simulation. Default value is 600.

bin_size::Float64: Size of the bin for pair correlation function and structure factor calculations. Default value is 0.05.

r_range::Float64: Range of r values for pair correlation function calculation. Default value is 10.

k_range::Float64: Range of k values for structure factor calculation. Default value is 10.

g2_weight_range::Float64: Weight range for pair correlation function in the objective function. Default value is 2.

s_weight_range::Float64: Weight range for structure factor in the objective function. Default value is 4.

n_threads::Int: Number of threads for parallel computation. Default value is 15.

configs_per_box::Int: Number of configurations per box for simulation. Default value is 10.

Ψ_tol::Float64: Tolerance for convergence of the objective function. Default value is 0.005.

show_pb::Bool: Boolean indicating whether to display a progress bar during simulation. Default value is true.

test::Bool: Boolean flag to indicate whether this is a test run and return a boolean indicating convergence. Default value is false.

Returns

If test is true, returns true if convergence is achieved, false otherwise.

If test is false, returns the optimized potential parameters.

Example

#form of the pair potential
pot(r, params) = params[1]*exp(-(r/params[2])^2)

#initial guess parameters
my_params = [1.0, 2.0] #write 1.0 instead of 1 to indicate that this is Float64

#target pair correlation function and structure factor
targ_g2(r) = 1 
targ_s(r) = 1

#optimize the parameters
Spec2Struc.optim_parametrized_pot(my_params, pot, 2, 1, targ_g2, targ_s)

Spec2Struc.reverse_mc — Function

reverse_mc(dim, n, ρ, g2_targ; initial_box = missing, bin_size = 0.05, range = 5, sweeps = 100, displace = 0.1, t_i = 1, t_f = 0.001, cooling_rate = 0.98)

Reverse Monte Carlo algorithm to generate equilibrium configurations that yield a target pair correlation function $g_2(r)$.

Arguments

dim::Int: Dimensionality of the system.

n::Int: Number of particles.

ρ::Float64: Number density of the particles.

g2_targ::Function: Target pair correlation function as a function g2_targ(r).

Keyword arguments

initial_box::Box (optional): Initial configuration box. Default is missing which generates a random box.

bin_size::Float64 (optional): Bin size for computing the pair correlation function. Default is 0.05.

range::Float64 (optional): Range for the interaction potential. Default is 5.

sweeps::Int (optional): Number of Monte Carlo sweeps at each temprature. Default is 100.

displace::Float64 (optional): Maximum displacement for particle moves. Default is 0.1.

t_i::Float64 (optional): Initial temperature. Default is 1.

t_f::Float64 (optional): Final temperature. Default is 0.001.

cooling_rate::Float64 (optional): Cooling rate for temperature reduction. Default is 0.98.

Returns

b::Box: Generated equilibrium classical configuration box. Use b.particles' to get the particle positions.

Example

box = Spec2Struc.reverse_mc(2, 100, 0.5, r -> 1 - exp(π*r^2))