Spec2Struc.jl

Efficient statistical mechanical algorithms to generate effective potentials or configurations from pair correlation functions or structure factors.

Package Features

Torquato-Wang algorithm (Torquato and Wang, PRE 2020)

Iterative Boltzmann Inversion (Soper, Chem. Phys. 1996)

Reverse Monte-Carlo (McGreevy, J. Phys. Cond. Matter 2001)

Function Documentation

Spec2Struc.optim_parametrized_pot — Function

optim_parametrized_pot(my_params; <keyword arguments>)

Optimize the parameters of a parametrized potential, starting with the initial guesses my_params

Arguments

pot: the functional form of the potential

large_r_grid = missing: ewald summation to treat very long-ranged interactions

dim: dimension of the system

n: number of particles

ρ: number density

bin_size: bin size for the pair correlation function g_2

r_range: the range of r in direct space to compute the pair correlation function g_2(r)

k_range: the range of k in fourier space to compute the structure factor S(k)

targ_g2: target pair correlation function

g2_weight_range: range parameter of the weight function for g2: wg2(r) = exp(-(r/g2weightrange)^2)

targ_s: target structure factor

s_weight_range: range parametrer of the weight function for S: ws(k) = exp(-(k/g2weight_range)^2)

n_threads: number of threads used for parallel computation of the simulation

configs_per_box: number of configurations to generate in each thread

Ψ_tol: stopping criterion for the error functional Ψ (distance between target and optimized pair statistics)

show_pb = true: show progress bar

test = false: whether we are testing the package